+++ tests/testthat/helper_learner.R
@@ -33,10 +33,9 @@ LearnerTorchTest1 = R6Class("LearnerTorchTest1",
       task_dataset(
         task,
         feature_ingress_tokens = list(num = ingress_token),
-        target_batchgetter = crate(function(data, device) {
-          torch_tensor(data = as.integer(data[[1]]), dtype = torch_long(), device = device)
-        }, .parent = topenv()),
-        device = param_vals$device
+        target_batchgetter = crate(function(data) {
+          torch_tensor(data = as.integer(data[[1]]), dtype = torch_long())
+        }, .parent = topenv())
       )
 
     },