+++ NEWS.md
@@ -23,6 +23,8 @@
 * feat: Added a callback to iteratively unfreeze parameters for finetuning.
 * fix: torch learners can now be used with `AutoTuner`.
 * feat: Added different learning rate schedulers as callbacks.
+* fix: `device` is no longer part of the `dataset` which allows for parallel dataloading
+  on GPUs.
 
 # mlr3torch 0.1.2